+++ mne/tests/test_docstring_parameters.py
@@ -59,6 +59,7 @@ def get_name(func):
 _deprecation_ignores = [
     'mne.io.write',  # always ignore these
     'mne.fixes._in1d',  # fix function
+    'mne.utils.plot_epochs_trellis'
 ]